Interactive symmetry copies

+++ src/bundles/phenix_ui/src/tool.py
@@ -40,6 +40,7 @@ class PhenixCitation(Citation):
         cite = '<br>'.join(["<b>" + title + "</b>"] + info)
         kw['prefix'] = "%s uses the Phenix <i>%s</i> command. Please cite:" % (tool_name, phenix_name)
         super().__init__(session, cite, **kw)
+
 class DouseSettings(Settings):
     AUTO_SAVE = {
         "show_hbonds": True,
@@ -203,6 +204,7 @@ class EmplaceLocalResultsViewer(ToolInstance):
 
         from chimerax.core.models import REMOVE_MODELS
         from chimerax.atomic import get_triggers
+        self._finalizing_symmetry = False
         self.handlers = [
             self.session.triggers.add_handler(REMOVE_MODELS, self._models_removed_cb),
         ]
@@ -261,10 +263,17 @@ class EmplaceLocalResultsViewer(ToolInstance):
     def accept_fit(self):
         # running commands directly in delete() can be hazardous, so...
         if self.sym_map and self.symmetry_checkbox.isChecked():
-            from chimerax.core.commands import run
+            self._finalizing_symmetry = True
+            from chimerax.core.commands import run, concise_model_spec, StringArg
             modelspec = self.orig_model.atomspec
+            prev_models = set(self.session.models[:])
             run(self.session,
                 f"sym clear {modelspec} ; sym {modelspec} symmetry {self.sym_map.atomspec} copies true")
+            added = [m for m in self.session.models if m not in prev_models]
+            orig_id, orig_name = self.orig_model.id[0], self.orig_model.name
+            run(self.session, f"close {modelspec}")
+            run(self.session, "combine " + concise_model_spec(self.session, added) + " close true"
+                " modelId %d name %s" % (orig_id, StringArg.unparse(orig_name)))
         self.delete()
 
     def delete(self):
@@ -331,10 +340,14 @@ class EmplaceLocalResultsViewer(ToolInstance):
             if finish:
                 tool._interpolate_handler.remove()
                 tool._interpolate_handler = None
+                if self.sym_map and self.symmetry_checkbox.isChecked():
+                    from chimerax.core.commands import run
+                    run(self.session, "sym " + self.orig_model.atomspec + " symmetry "
+                        + self.sym_map.atomspec + " copies false", log=False)
         self._interpolate_handler = self.session.triggers.add_handler("new frame", make_step)
 
     def _models_removed_cb(self, trig_name, trig_data):
-        if self.orig_model in trig_data:
+        if self.orig_model in trig_data and not self._finalizing_symmetry:
             self.delete()
 
     def _new_selection(self, selected, unselected):
